+++ contracts/test/phased/PhasedScheduleTest.sol
@@ -1,27 +1,26 @@
 // SPDX-License-Identifier: CAL
 pragma solidity ^0.8.10;
 
-import { Phase, Phased } from "../../phased/Phased.sol";
+import {Phase, Phased} from "../../phased/Phased.sol";
 
 /// @title PhasedScheduleTest
 /// Contract for testing phase hook functionality.
 contract PhasedScheduleTest is Phased {
-
     constructor() {
         initializePhased();
     }
 
     /// Exposes `scheduleNextPhase` for testing.
     /// @param nextPhaseBlock_ As per `scheduleNestPhase`.
-    function testScheduleNextPhase(uint nextPhaseBlock_) external {
+    function testScheduleNextPhase(uint256 nextPhaseBlock_) external {
         scheduleNextPhase(nextPhaseBlock_);
     }
 
-    function succeedsOnlyPhaseZero() internal onlyPhase(Phase.ZERO)
-    { } // solhint-disable-line no-empty-blocks
+    // solhint-disable-next-line no-empty-blocks
+    function succeedsOnlyPhaseZero() internal onlyPhase(Phase.ZERO) {}
 
     /// @inheritdoc Phased
-    function _beforeScheduleNextPhase(uint nextPhaseBlock_)
+    function _beforeScheduleNextPhase(uint256 nextPhaseBlock_)
         internal
         virtual
         override
